dynamic settings
for the case when elements exist it set high and low values to be able to cature the elements extents, when they do not then uses defaults for the borehole. When elements do not exist then it uses the borehole as the element and sets the current water level rect to use borehole values.